First commit
Spent some Time already working on
idea is to create a tower defense
i created a script where the spawned unit walks to a point of interest then turns and heads for the objective.
i created the spawn script and the move script
next steps will be to create the objective to protect and what happens when the enemy gets to it.